review fixes, move to indexing_optimizer.rs
