+++ lib/collection/src/collection_manager/optimizers/indexing_optimizer.rs
@@ -1124,4 +1124,117 @@ mod tests {
             "an HNSW index should have been created to promote the deferred points",
         );
     }
+
+    /// Non-appendable segments below the HNSW full-scan boundary are
+    /// scanned in full by every search, so the indexing optimizer must keep
+    /// merging them away even when the segment-count target is already
+    /// reached.
+    #[test]
+    fn test_merge_sub_full_scan_tails_at_target() {
+        let dir = Builder::new().prefix("segment_dir").tempdir().unwrap();
+        let temp_dir = Builder::new().prefix("segment_temp_dir").tempdir().unwrap();
+        let dim = 256;
+
+        let make_frozen = |num_vectors: u64| {
+            let mut segment = random_segment(dir.path(), 100, num_vectors, dim);
+            // The planner consults `Segment::is_appendable`, which reads the
+            // runtime flag — mark the segment as frozen directly.
+            segment.appendable_flag = false;
+            segment
+        };
+
+        // full_scan_threshold = 10 KB; dim=256 => 1 KB per vector.
+        // Tails: 3 vectors = 3 KB (below); big: 20 vectors = 20 KB (above).
+        let build_optimizer = || {
+            let collection_params = CollectionParams {
+                vectors: VectorsConfig::Single(
+                    VectorParamsBuilder::new(dim as u64, Distance::Dot).build(),
+                ),
+                ..CollectionParams::empty()
+            };
+            let hnsw_config = HnswConfig {
+                full_scan_threshold: 10,
+                ..Default::default()
+            };
+            new_indexing_optimizer(
+                5,
+                OptimizerThresholds {
+                    max_segment_size_kb: 100_000,
+                    memmap_threshold_kb: 1_000_000,
+                    indexing_threshold_kb: 1_000_000,
+                    deferred_internal_id: None,
+                },
+                dir.path().to_owned(),
+                temp_dir.path().to_owned(),
+                collection_params,
+                hnsw_config,
+                HnswGlobalConfig::default(),
+                None,
+            )
+        };
+
+        // Case 1: frozen (non-appendable) tails at the target count are merged
+        // together with the smallest regular segment.
+        let mut holder = SegmentHolder::default();
+        let tails = [
+            holder.add_new(make_frozen(3)),
+            holder.add_new(make_frozen(3)),
+            holder.add_new(make_frozen(3)),
+        ];
+        let big = [
+            holder.add_new(make_frozen(20)),
+            holder.add_new(make_frozen(20)),
+        ];
+        let locked_holder = LockedSegmentHolder::new(holder);
+        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
+        let batch = plans.into_iter().exactly_one().unwrap();
+        assert_eq!(batch.len(), 4, "3 tails + 1 neighbour expected");
+        for tail in &tails {
+            assert!(batch.contains(tail), "tail must be merged");
+        }
+        assert_eq!(
+            batch.iter().filter(|sid| big.contains(sid)).count(),
+            1,
+            "exactly one regular segment joins the merge"
+        );
+
+        // Case 2: appendable segments of the same sizes are left alone — they
+        // are still receiving writes and must not be treated as tails.
+        let mut holder = SegmentHolder::default();
+        for num_vectors in [3, 3, 3, 20, 20] {
+            holder.add_new(random_segment(dir.path(), 100, num_vectors, dim));
+        }
+        let locked_holder = LockedSegmentHolder::new(holder);
+        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
+        assert!(plans.is_empty(), "appendable segments must not be planned");
+
+        // Case 3: no partner above the boundary and the tails together stay
+        // below it — merging cannot cross the boundary, so nothing is planned
+        // (the segment count must not be reduced for no gain).
+        let mut holder = SegmentHolder::default();
+        holder.add_new(make_frozen(3));
+        holder.add_new(make_frozen(3));
+        let locked_holder = LockedSegmentHolder::new(holder);
+        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
+        assert!(
+            plans.is_empty(),
+            "a merge that cannot cross the full-scan boundary must not be planned"
+        );
+
+        // Case 4: no partner, but the tails add up past the boundary — they
+        // are merged on their own.
+        let mut holder = SegmentHolder::default();
+        let tails = [
+            holder.add_new(make_frozen(6)),
+            holder.add_new(make_frozen(6)),
+        ];
+        let locked_holder = LockedSegmentHolder::new(holder);
+        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
+        let batch = plans.into_iter().exactly_one().unwrap();
+        assert_eq!(
+            batch.iter().sorted().collect_vec(),
+            tails.iter().sorted().collect_vec(),
+            "tails crossing the boundary together must be merged without a partner"
+        );
+    }
 }
