+++ lib/collection/src/collection_manager/optimizers/merge_optimizer.rs
@@ -286,95 +286,4 @@ mod tests {
             );
         }
     }
-
-    /// Non-appendable segments below the HNSW full-scan boundary are
-    /// scanned in full by every search, so the merge optimizer must keep
-    /// merging them even when the segment-count target is already reached.
-    #[test]
-    fn test_merge_sub_full_scan_tails_at_target() {
-        use crate::config::CollectionParams;
-        use crate::operations::types::VectorsConfig;
-        use crate::operations::vector_params_builder::VectorParamsBuilder;
-        use crate::optimizers_builder::build_segment_optimizer_config;
-
-        let dir = Builder::new().prefix("segment_dir").tempdir().unwrap();
-        let temp_dir = Builder::new().prefix("segment_temp_dir").tempdir().unwrap();
-        let dim = 256;
-
-        let make_frozen = |num_vectors: u64| {
-            let mut segment = random_segment(dir.path(), 100, num_vectors, dim);
-            // The planner consults `Segment::is_appendable`, which reads the
-            // runtime flag — mark the segment as frozen directly.
-            segment.appendable_flag = false;
-            segment
-        };
-
-        // full_scan_threshold = 10 KB; dim=256 => 1 KB per vector.
-        // Tails: 3 vectors = 3 KB (below); big: 20 vectors = 20 KB (above).
-        let build_optimizer = || {
-            let collection_params = CollectionParams {
-                vectors: VectorsConfig::Single(
-                    VectorParamsBuilder::new(dim as u64, Distance::Dot).build(),
-                ),
-                ..CollectionParams::empty()
-            };
-            let hnsw_config = HnswConfig {
-                full_scan_threshold: 10,
-                ..Default::default()
-            };
-            let segment_config = build_segment_optimizer_config(
-                &collection_params,
-                &hnsw_config,
-                &Default::default(),
-            );
-            MergeOptimizer::new(
-                5,
-                OptimizerThresholds {
-                    max_segment_size_kb: 100_000,
-                    memmap_threshold_kb: 1_000_000,
-                    indexing_threshold_kb: 1_000_000,
-                    deferred_internal_id: None,
-                },
-                dir.path().to_owned(),
-                temp_dir.path().to_owned(),
-                segment_config,
-                HnswGlobalConfig::default(),
-            )
-        };
-
-        // Case 1: frozen (non-appendable) tails at the target count are merged
-        // together with the smallest regular segment.
-        let mut holder = SegmentHolder::default();
-        let tails = [
-            holder.add_new(make_frozen(3)),
-            holder.add_new(make_frozen(3)),
-            holder.add_new(make_frozen(3)),
-        ];
-        let big = [
-            holder.add_new(make_frozen(20)),
-            holder.add_new(make_frozen(20)),
-        ];
-        let locked_holder = LockedSegmentHolder::new(holder);
-        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
-        let batch = plans.into_iter().exactly_one().unwrap();
-        assert_eq!(batch.len(), 4, "3 tails + 1 neighbour expected");
-        for tail in &tails {
-            assert!(batch.contains(tail), "tail must be merged");
-        }
-        assert_eq!(
-            batch.iter().filter(|sid| big.contains(sid)).count(),
-            1,
-            "exactly one regular segment joins the merge"
-        );
-
-        // Case 2: appendable segments of the same sizes are left alone — they
-        // are still receiving writes and must not be treated as tails.
-        let mut holder = SegmentHolder::default();
-        for num_vectors in [3, 3, 3, 20, 20] {
-            holder.add_new(random_segment(dir.path(), 100, num_vectors, dim));
-        }
-        let locked_holder = LockedSegmentHolder::new(holder);
-        let plans = build_optimizer().plan_optimizations_for_test(&locked_holder);
-        assert!(plans.is_empty(), "appendable segments must not be planned");
-    }
 }
